Consolidate agentic development workflows Do not use API key for Gemini.
