--- docker/Dockerfile.claude
@@ -1,41 +1,0 @@
-# Dockerfile for Ubuntu 24.04 with Claude Code in YOLO mode.
-FROM ubuntu:24.04
-
-# Install supporting tools for downloading/installing later tools.
-RUN apt-get update && apt-get install -y --no-install-recommends \
-    curl \
-    git \
-    python-is-python3 \
-    python3-pip \
-    python3-venv \
-    unzip \
-    wget \
-&& rm -rf /var/lib/apt/lists/*
-
-# Create non-root user
-RUN useradd -m -s /bin/bash vscode \
-    && mkdir -p /workspace \
-    && chown vscode:vscode /workspace
-
-# Install Node.js (required for Claude Code)
-RUN curl -fsSL https://deb.nodesource.com/setup_20.x | bash - \
-    && apt-get install -y --no-install-recommends nodejs \
-    && rm -rf /var/lib/apt/lists/*
-
-# Switch to non-root user
-USER vscode
-WORKDIR /workspace
-
-# Configure npm for user-local global installs
-ENV NPM_CONFIG_PREFIX=/home/vscode/.npm-global
-ENV PATH="/home/vscode/.npm-global/bin:/home/vscode/.local/bin:/usr/local/bin:${PATH}"
-RUN mkdir -p /home/vscode/.npm-global
-
-# Install Claude Code as the user
-RUN npm install -g @anthropic-ai/claude-code
-
-# Set up uv and Python environment
-RUN curl -LsSf https://astral.sh/uv/install.sh | sh
-ENV UV_PROJECT_ENVIRONMENT="/home/vscode/.venv"
-
-# ~/.claude is mounted from the host at runtime to persist auth state between sessions.
