--- scripts/claude-sandbox.sh
@@ -1,10 +1,0 @@
-#!/bin/bash
-
-set -eu -o pipefail
-
-# Get the workspace root
-WORKSPACE_ROOT=$(git rev-parse --show-toplevel)
-
-# Change directory to the workspace root and run the python script
-cd "$WORKSPACE_ROOT"
-uv run scripts/claude-sandbox.py "$@"
